VideoToVideo process prototype

+++ OnnxStack.Core/Video/VideoInfo.cs
@@ -2,5 +2,5 @@
 
 namespace OnnxStack.Core.Video
 {
-    public record VideoInfo(int Width, int Height, TimeSpan Duration, int FPS);
+    public record VideoInfo(int Width, int Height, TimeSpan Duration, float FPS);
 }
